refactor slide puzzles

+++ src/slide_puzzle.h
@@ -13,9 +13,9 @@ protected:
 	static void _bind_methods();
 
 public:
-	static PackedVector2Array shuffle(int p_complexity, Array p_squares, int p_moves, const Ref<RandomNumberGenerator> &p_rng);
-	static bool is_solvable(int p_complexity, const PackedInt32Array &p_squares);
-	static PackedVector2Array solve(int p_complexity, const PackedInt32Array &p_squares, const PackedInt32Array &p_goal = {});
+	static PackedVector2Array shuffle(uint8_t p_complexity, Array p_squares, int64_t p_moves, const Ref<RandomNumberGenerator> &p_rng);
+	static bool is_solvable(uint8_t p_complexity, const PackedInt32Array &p_squares);
+	static PackedVector2Array solve(uint8_t p_complexity, const PackedInt32Array &p_squares, const PackedInt32Array &p_goal = {});
 };
 
 } //namespace godot
